+++ src/aarch64/cpu-aarch64.cc
@@ -66,7 +66,7 @@ uint32_t CPU::GetCacheType() {
   __asm__ __volatile__("mrs %[ctr], ctr_el0"  // NOLINT
                        : [ctr] "=r"(cache_type_register));
   VIXL_ASSERT(IsUint32(cache_type_register));
-  return cache_type_register;
+  return static_cast<uint32_t>(cache_type_register);
 #else
   // This will lead to a cache with 1 byte long lines, which is fine since
   // neither EnsureIAndDCacheCoherency nor the simulator will need this
